Added support for pipe transforms to directly write into stdin, instead of using a temporary file.
svn path=/plone.transforms/trunk/; revision=16309

+++ plone/transforms/pipe.py
@@ -3,7 +3,7 @@ import tempfile
 
 from zope.interface import implements
 
-from plone.transforms.interfaces import ICommandTransform
+from plone.transforms.interfaces import IPipeTransform
 
 from plone.transforms.transform import PersistentTransform
 from plone.transforms.transform import TransformResult
@@ -18,11 +18,11 @@ class PipeTransform(PersistentTransform):
     Let's make sure that this implementation actually fulfills the API.
 
       >>> from zope.interface.verify import verifyClass
-      >>> verifyClass(ICommandTransform, PipeTransform)
+      >>> verifyClass(IPipeTransform, PipeTransform)
       True
     """
 
-    implements(ICommandTransform)
+    implements(IPipeTransform)
 
     inputs = (None, )
     output = None
@@ -34,6 +34,15 @@ class PipeTransform(PersistentTransform):
 
     command = None
     args = None
+    use_stdin = False
+
+    def write(self, fd, data):
+        # write data to tmp using a file descriptor
+        firstchunk = data.next()
+        if isinstance(firstchunk, unicode):
+            self.writeText(fd, firstchunk, data)
+        else:
+            self.writeBinary(fd, firstchunk, data)
 
     def writeBinary(self, fd, firstchunk, data):
         os.write(fd, firstchunk)
@@ -49,15 +58,11 @@ class PipeTransform(PersistentTransform):
         """Create a temporary directory, copy input in a file there
         return the path of the tmp dir and of the input file.
         """
-        filehandle, tmpname = tempfile.mkstemp(text=False)
+        fd, tmpname = tempfile.mkstemp(text=False)
         # write data to tmp using a file descriptor
-        firstchunk = data.next()
-        if isinstance(firstchunk, unicode):
-            self.writeText(filehandle, firstchunk, data)
-        else:
-            self.writeBinary(filehandle, firstchunk, data)
+        self.write(fd, data)
         # close it so the other process can read it
-        os.close(filehandle)
+        os.close(fd)
         return tmpname
 
     def extractOutput(self, stdout):
@@ -68,17 +73,21 @@ class PipeTransform(PersistentTransform):
         The transform method takes some data in one of the input formats and
         returns it in the output format.
         """
-        tmpname = self.initialize_tmpfile(data)
-        # do some stuff
-        commandline = "%s %s" % (self.command, self.args)
-        commandline = commandline % { 'infile' : tmpname }
+        if not self.use_stdin:
+            tmpname = self.initialize_tmpfile(data)
+            commandline = "%s %s" % (self.command, self.args)
+            commandline = commandline % { 'infile' : tmpname }
 
         child_stdin, child_stdout = os.popen4(commandline, 'b')
+
+        if self.use_stdin:
+            self.write(child_stdin, data)
 
         status = child_stdin.close()
         out = self.extractOutput(child_stdout)
         child_stdout.close()
 
-        os.unlink(tmpname)
+        if not self.use_stdin:
+            os.unlink(tmpname)
 
         return TransformResult(iter(out))
